chore: clean up refactor and comments.
