+++ supabase/functions/_utils/stripe.ts
@@ -1,0 +1,9 @@
+// esm.sh is used to compile stripe-node to be compatible with ES modules.
+import Stripe from "https://esm.sh/stripe?target=deno&no-check";
+
+export const stripe = Stripe(Deno.env.get("STRIPE_SECRET_KEY") ?? "", {
+  // This is needed to use the Fetch API rather than relying on the Node http
+  // package.
+  httpClient: Stripe.createFetchHttpClient(),
+  apiVersion: "2020-08-27",
+});
